+++ txmatching/scorers/compatibility_graph.py
@@ -1,3 +1,3 @@
 from typing import Dict, Tuple
 
-CompatibilityGraph = Dict[Tuple[int, int], float]
+CompatibilityGraph = Dict[Tuple[int, int], int]
